refactor(command): move the command bar panel out of the layout page
This is the bar people actually see. `handle_open_command_bar` emits to
`LAYOUT_COMMAND_BAR_OPEN_EVENT`, so Cmd+K opens `CommandBarPanel` inside the
layout page's own DOM, not the `vmux://command-bar/` webview - which meant the
window shell was the implementer of the surface, not merely its host.

The component moves to `vmux_command::panel` with everything it needs: the drag
and resize machinery, `PanelPlacement`, `clamp_panel_placement`, the two panel
events and the two channel names, plus the three clamp tests.

It also stops being three props and two listeners the host page has to hold. The
panel owns its own state now and returns nothing when closed, so the layout page
renders `CommandBarPanel {}` and knows nothing else about it - no signals, no
open flag, no placement, no `set_command_bar_panel_active`.

Drag and resize deliberately stay in the page: they move a DOM node, so routing
them through the ECS would put an IPC round trip and a Bevy frame between the
pointer and the pixels.
