+++ crates/page/vmux_layout/src/event.rs
@@ -6,102 +6,6 @@ pub const SERVICES_PAGE_URL: &str = "vmux://services/";
 pub const LAYOUT_STATE_EVENT: &str = "layout-state";
 pub const STACKS_EVENT: &str = "stacks";
 pub const RELOAD_EVENT: &str = "reload";
-/// Host -> layout page: open the command bar panel with a fresh payload. Distinct from
-/// `COMMAND_BAR_OPEN_EVENT` so the layout page and the start page can be addressed separately.
-pub const LAYOUT_COMMAND_BAR_OPEN_EVENT: &str = "layout-command-bar-open";
-
-/// Host -> layout page: close the command bar panel.
-///
-/// The host cannot close the panel by mutating a surface the way the modal allowed, so dismiss
-/// and the `Cmd+K`-while-open toggle both have to ask the page to unmount it.
-pub const LAYOUT_COMMAND_BAR_CLOSE_EVENT: &str = "layout-command-bar-close";
-
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    serde::Serialize,
-    serde::Deserialize,
-    rkyv::Archive,
-    rkyv::Serialize,
-    rkyv::Deserialize,
-)]
-pub struct CommandBarPanelCloseEvent;
-
-/// Where the user has dragged and sized the floating command bar, in CSS pixels.
-///
-/// Absent until the first drag or resize, so an untouched bar keeps its centred default and
-/// content-driven height.
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    PartialEq,
-    serde::Serialize,
-    serde::Deserialize,
-    rkyv::Archive,
-    rkyv::Serialize,
-    rkyv::Deserialize,
-)]
-pub struct PanelPlacement {
-    pub left: f64,
-    pub top: f64,
-    pub width: f64,
-    pub height: f64,
-}
-
-pub const PANEL_MIN_WIDTH: f64 = 320.0;
-pub const PANEL_MIN_HEIGHT: f64 = 120.0;
-
-/// Keep the bar inside the window and above a usable minimum.
-///
-/// Without the clamp a drag can park the bar past the edge, where it is unreachable and there is
-/// no chrome to drag it back by.
-pub fn clamp_panel_placement(
-    placement: PanelPlacement,
-    viewport_width: f64,
-    viewport_height: f64,
-) -> PanelPlacement {
-    let width = placement
-        .width
-        .clamp(PANEL_MIN_WIDTH, viewport_width.max(PANEL_MIN_WIDTH));
-    let height = placement
-        .height
-        .clamp(PANEL_MIN_HEIGHT, viewport_height.max(PANEL_MIN_HEIGHT));
-    PanelPlacement {
-        left: placement.left.clamp(0.0, (viewport_width - width).max(0.0)),
-        top: placement
-            .top
-            .clamp(0.0, (viewport_height - height).max(0.0)),
-        width,
-        height,
-    }
-}
-
-/// Layout page -> host: the command bar panel took or released the keyboard.
-///
-/// Mirrors `BookmarkTextInputEvent`: while the panel holds a focused DOM field the layout shell
-/// must own `CefKeyboardTarget`, or keystrokes go to the focused pane instead.
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    Default,
-    PartialEq,
-    Eq,
-    serde::Serialize,
-    serde::Deserialize,
-    rkyv::Archive,
-    rkyv::Serialize,
-    rkyv::Deserialize,
-)]
-pub struct CommandBarPanelActiveEvent {
-    pub active: bool,
-}
-
 #[derive(
     Clone,
     Copy,
@@ -251,71 +155,6 @@ pub const SIDE_SHEET_WIDTH_PX: f32 = 220.0;
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    /// A bar dragged past the edge has no chrome left to grab, so the clamp is the only thing
-    /// keeping it recoverable.
-    #[test]
-    fn dragging_the_panel_off_screen_keeps_it_reachable() {
-        let dragged_past_the_corner = PanelPlacement {
-            left: 5000.0,
-            top: 5000.0,
-            width: 576.0,
-            height: 400.0,
-        };
-
-        let clamped = clamp_panel_placement(dragged_past_the_corner, 1440.0, 900.0);
-
-        assert_eq!(clamped.left, 1440.0 - 576.0);
-        assert_eq!(clamped.top, 900.0 - 400.0);
-        assert_eq!(clamped.width, 576.0);
-        assert_eq!(clamped.height, 400.0);
-
-        let dragged_past_the_origin = PanelPlacement {
-            left: -300.0,
-            top: -80.0,
-            ..dragged_past_the_corner
-        };
-
-        let clamped = clamp_panel_placement(dragged_past_the_origin, 1440.0, 900.0);
-
-        assert_eq!(clamped.left, 0.0);
-        assert_eq!(clamped.top, 0.0);
-    }
-
-    /// Resizing below the minimum collapses the bar to a sliver with no visible resize handle.
-    #[test]
-    fn resizing_the_panel_stops_at_the_minimum() {
-        let collapsed = PanelPlacement {
-            left: 10.0,
-            top: 10.0,
-            width: 10.0,
-            height: 10.0,
-        };
-
-        let clamped = clamp_panel_placement(collapsed, 1440.0, 900.0);
-
-        assert_eq!(clamped.width, PANEL_MIN_WIDTH);
-        assert_eq!(clamped.height, PANEL_MIN_HEIGHT);
-    }
-
-    /// A window narrower than the minimum must still produce a placement inside it rather than a
-    /// negative offset.
-    #[test]
-    fn panel_survives_a_window_smaller_than_its_minimum() {
-        let placement = PanelPlacement {
-            left: 40.0,
-            top: 40.0,
-            width: 576.0,
-            height: 400.0,
-        };
-
-        let clamped = clamp_panel_placement(placement, 200.0, 100.0);
-
-        assert_eq!(clamped.left, 0.0);
-        assert_eq!(clamped.top, 0.0);
-        assert_eq!(clamped.width, PANEL_MIN_WIDTH);
-        assert_eq!(clamped.height, PANEL_MIN_HEIGHT);
-    }
 
     #[test]
     fn main_cef_left_includes_side_sheet_gap_when_open() {
