+++ crates/vmux_browser/src/command_bar/panel.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 use bevy_cef::prelude::{BinEventEmitterPlugin, BinReceive};
 
-use vmux_layout::event::CommandBarPanelActiveEvent;
+use vmux_command::event::CommandBarPanelActiveEvent;
 
 pub struct CommandBarPanelPlugin;
